+++ src/lib/app-provider/hooks/index.ts
@@ -15,3 +15,4 @@ export * from "./useCurrentChain";
 export * from "./useConvertHexAddress";
 export * from "./usePlatform";
 export * from "./useInitia";
+export * from "./useIsConnected";