+++ src/lib/components/InputWithIcon.tsx
@@ -1,66 +1,34 @@
 import type { InputProps } from "@chakra-ui/react";
-import {
-  InputLeftElement,
-  Input,
-  InputGroup,
-  InputRightElement,
-} from "@chakra-ui/react";
-import type { ChangeEvent } from "react";
+import { InputLeftElement, Input, InputGroup } from "@chakra-ui/react";
 
 import { AmpEvent, track } from "lib/amplitude";
 
 import { CustomIcon } from "./icon";
 
-interface InputWithIconProps {
-  placeholder: string;
-  value: string;
-  size?: InputProps["size"];
-  my?: InputProps["my"];
-  autoFocus?: boolean;
-  action?: string;
-  iconPosition?: "start" | "end";
-  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+interface InputWithIconProps extends InputProps {
+  amptrackSection?: string;
 }
 
-const SearchIcon = () => <CustomIcon name="search" color="gray.600" />;
-
 const InputWithIcon = ({
-  placeholder,
-  value,
-  size,
   my,
-  action,
-  iconPosition = "end",
-  autoFocus = false,
-  onChange,
+  size = "md",
+  amptrackSection,
+  ...inputProps
 }: InputWithIconProps) => (
   <InputGroup my={my}>
+    <InputLeftElement h="full" alignItems="center">
+      <CustomIcon name="search" color="gray.600" />
+    </InputLeftElement>
     <Input
-      placeholder={placeholder}
-      value={value}
+      {...inputProps}
       size={size}
-      p={`8px ${iconPosition === "end" ? "40px" : "12px"} 8px ${
-        iconPosition === "start" ? "40px" : "12px"
-      } !important`}
-      onChange={onChange}
-      autoFocus={autoFocus}
+      paddingLeft="36px !important"
       onClick={
-        action ? () => track(AmpEvent.USE_SEARCH_INPUT, { action }) : undefined
+        amptrackSection
+          ? () => track(AmpEvent.USE_SEARCH_INPUT, { amptrackSection })
+          : undefined
       }
     />
-    {iconPosition === "end" ? (
-      <InputRightElement
-        h={size === "lg" ? "56px" : "full"}
-        alignItems="center"
-        mr={1}
-      >
-        <SearchIcon />
-      </InputRightElement>
-    ) : (
-      <InputLeftElement h={size === "lg" ? "56px" : "full"} alignItems="center">
-        <SearchIcon />
-      </InputLeftElement>
-    )}
   </InputGroup>
 );