+++ src/lib/layout/Searchbar.tsx
@@ -1,7 +1,4 @@
 import {
-  Input,
-  InputGroup,
-  InputRightElement,
   ListItem,
   List,
   FormControl,
@@ -29,6 +26,7 @@ import {
   useMobile,
 } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
+import InputWithIcon from "lib/components/InputWithIcon";
 import { PrimaryNameMark } from "lib/components/PrimaryNameMark";
 import type {
   ResultMetadata,
@@ -225,14 +223,17 @@ const getNextCursor = (
 const getPlaceholder = ({
   isWasm,
   isPool,
+  isMove,
 }: {
   isWasm: boolean;
   isPool: boolean;
+  isMove: boolean;
 }) => {
-  const wasmText = isWasm ? "/ Code ID / Contract Address" : "";
-  const poolText = isPool ? "/ Pool ID" : "";
+  const wasmText = isWasm ? "/ Code ID / Contract Address " : "";
+  const poolText = isPool ? "/ Pool ID " : "";
+  const moveText = isMove ? "/ Module Path " : "";
 
-  return `Search by Account Address / Tx Hash / Block ${wasmText} ${poolText}`;
+  return `Search by Account Address / Tx Hash / Block${wasmText}${poolText}${moveText}`;
 };
 
 // eslint-disable-next-line sonarjs/cognitive-complexity
@@ -247,6 +248,7 @@ const Searchbar = () => {
       features: {
         wasm: { enabled: isWasm },
         pool: { enabled: isPool },
+        move: { enabled: isMove },
       },
     },
   } = useCelatoneApp();
@@ -325,7 +327,7 @@ const Searchbar = () => {
         <DrawerContent>
           <DrawerBody overflowY="scroll" p={2} m={2}>
             <FormControl ref={boxRef}>
-              <InputGroup mb={4} alignItems="center">
+              <Flex mb={4}>
                 <IconButton
                   fontSize="24px"
                   variant="gray"
@@ -334,21 +336,16 @@ const Searchbar = () => {
                   color="gray.600"
                   icon={<CustomIcon name="chevron-left" />}
                 />
-                <Input
+                <InputWithIcon
                   value={keyword}
-                  h="36px"
                   onChange={handleSearchChange}
                   placeholder="Type your keyword ..."
-                  focusBorderColor="secondary.main"
                   autoFocus
                   onFocus={() => setDisplayResults(keyword.length > 0)}
                   onKeyDown={(e) => handleOnKeyEnter(e, onClose)}
                   autoComplete="off"
                 />
-                <InputRightElement pointerEvents="none" h="full">
-                  <CustomIcon name="search" color="gray.600" />
-                </InputRightElement>
-              </InputGroup>
+              </Flex>
               {displayResults ? (
                 <List borderRadius="8px" bg="gray.900" w="full">
                   {isLoading || isTyping ? (
@@ -393,7 +390,7 @@ const Searchbar = () => {
               )}
             </FormControl>
             <Text variant="body3" color="text.dark" textAlign="center" mt={2}>
-              {getPlaceholder({ isWasm, isPool })}
+              {getPlaceholder({ isWasm, isPool, isMove })}
             </Text>
           </DrawerBody>
         </DrawerContent>
@@ -401,21 +398,14 @@ const Searchbar = () => {
     </>
   ) : (
     <FormControl ref={boxRef} zIndex={3}>
-      <InputGroup>
-        <Input
-          value={keyword}
-          h="36px"
-          onChange={handleSearchChange}
-          placeholder={getPlaceholder({ isWasm, isPool })}
-          focusBorderColor="secondary.main"
-          onFocus={() => setDisplayResults(keyword.length > 0)}
-          onKeyDown={handleOnKeyEnter}
-          autoComplete="off"
-        />
-        <InputRightElement pointerEvents="none" h="full">
-          <CustomIcon name="search" color="gray.600" />
-        </InputRightElement>
-      </InputGroup>
+      <InputWithIcon
+        value={keyword}
+        onChange={handleSearchChange}
+        placeholder={getPlaceholder({ isWasm, isPool, isMove })}
+        onFocus={() => setDisplayResults(keyword.length > 0)}
+        onKeyDown={handleOnKeyEnter}
+        autoComplete="off"
+      />
       {displayResults && (
         <List
           borderRadius="8px"