+++ src/lib/pages/account-details/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable complexity */
 import {
   Flex,
   TabList,
@@ -27,7 +28,7 @@ import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
 import { useAccountInfo } from "lib/services/accountService";
 import { useModulesByAddress } from "lib/services/move/moduleService";
 import { useResourcesByAddress } from "lib/services/move/resourceService";
-import type { Addr, HexAddr, HumanAddr } from "lib/types";
+import type { Addr, HexAddr, HumanAddr, Option } from "lib/types";
 import { truncate } from "lib/utils";
 
 import { AccountHeader } from "./components/AccountHeader";
@@ -98,9 +99,9 @@ const AccountDetailsBody = ({
   // -----------------CALLBACKS----------------//
   // ------------------------------------------//
   const handleTabChange = useCallback(
-    (nextTab: TabIndex) => () => {
+    (nextTab: TabIndex, total: Option<number>) => () => {
       if (nextTab === tabParam) return;
-      trackUseTab(nextTab);
+      trackUseTab(nextTab, undefined, total?.toString());
       navigate({
         pathname: "/accounts/[accountAddress]/[tab]",
         query: {
@@ -149,18 +150,18 @@ const AccountDetailsBody = ({
           overflowX="scroll"
           id={tableHeaderId}
         >
-          <CustomTab onClick={handleTabChange(TabIndex.Overview)}>
+          <CustomTab onClick={handleTabChange(TabIndex.Overview, undefined)}>
             Overview
           </CustomTab>
           <CustomTab
             count={tableCounts.assetsCount}
             isDisabled={tableCounts.assetsCount === 0}
-            onClick={handleTabChange(TabIndex.Assets)}
+            onClick={handleTabChange(TabIndex.Assets, undefined)}
           >
             Assets
           </CustomTab>
           <CustomTab
-            onClick={handleTabChange(TabIndex.Delegations)}
+            onClick={handleTabChange(TabIndex.Delegations, undefined)}
             hidden={disableDelegation}
           >
             Delegations
@@ -168,7 +169,10 @@ const AccountDetailsBody = ({
           <CustomTab
             count={tableCounts.txsCount}
             isDisabled={tableCounts.txsCount === 0}
-            onClick={handleTabChange(TabIndex.Txs)}
+            onClick={handleTabChange(
+              TabIndex.Txs,
+              tableCounts.txsCount ?? undefined
+            )}
             isLoading={isLoadingAccountTableCounts}
           >
             Transactions
@@ -176,7 +180,10 @@ const AccountDetailsBody = ({
           <CustomTab
             count={tableCounts.codesCount}
             isDisabled={tableCounts.codesCount === 0}
-            onClick={handleTabChange(TabIndex.Codes)}
+            onClick={handleTabChange(
+              TabIndex.Codes,
+              tableCounts.codesCount ?? undefined
+            )}
             isLoading={isLoadingAccountTableCounts}
             hidden={!wasm.enabled}
           >
@@ -185,7 +192,10 @@ const AccountDetailsBody = ({
           <CustomTab
             count={tableCounts.contractsCount}
             isDisabled={tableCounts.contractsCount === 0}
-            onClick={handleTabChange(TabIndex.Contracts)}
+            onClick={handleTabChange(
+              TabIndex.Contracts,
+              tableCounts.contractsCount ?? undefined
+            )}
             isLoading={isLoadingAccountTableCounts}
             hidden={!wasm.enabled}
           >
@@ -194,7 +204,10 @@ const AccountDetailsBody = ({
           <CustomTab
             count={tableCounts.contractsAdminCount}
             isDisabled={tableCounts.contractsAdminCount === 0}
-            onClick={handleTabChange(TabIndex.Admins)}
+            onClick={handleTabChange(
+              TabIndex.Admins,
+              tableCounts.contractsAdminCount ?? undefined
+            )}
             isLoading={isLoadingAccountTableCounts}
             hidden={!wasm.enabled}
           >
@@ -203,7 +216,10 @@ const AccountDetailsBody = ({
           <CustomTab
             count={resourcesData?.totalCount}
             isDisabled={resourcesData?.totalCount === 0}
-            onClick={handleTabChange(TabIndex.Resources)}
+            onClick={handleTabChange(
+              TabIndex.Resources,
+              resourcesData?.groupedByOwner.length
+            )}
             hidden={!move.enabled}
           >
             Resources
@@ -211,7 +227,7 @@ const AccountDetailsBody = ({
           <CustomTab
             count={modulesData?.length}
             isDisabled={modulesData?.length === 0}
-            onClick={handleTabChange(TabIndex.Modules)}
+            onClick={handleTabChange(TabIndex.Modules, undefined)}
             hidden={!move.enabled}
           >
             Modules
@@ -220,7 +236,10 @@ const AccountDetailsBody = ({
             count={tableCounts.proposalsCount}
             isDisabled={tableCounts.proposalsCount === 0}
             isLoading={isLoadingAccountTableCounts}
-            onClick={handleTabChange(TabIndex.Proposals)}
+            onClick={handleTabChange(
+              TabIndex.Proposals,
+              tableCounts.proposalsCount ?? undefined
+            )}
             hidden={!gov.enabled}
           >
             Proposals
@@ -262,7 +281,7 @@ const AccountDetailsBody = ({
             >
               <AssetsSection
                 address={accountAddress}
-                onViewMore={handleTabChange(TabIndex.Assets)}
+                onViewMore={handleTabChange(TabIndex.Assets, undefined)}
               />
             </Flex>
             {!disableDelegation && (
@@ -272,7 +291,7 @@ const AccountDetailsBody = ({
               >
                 <DelegationsSection
                   walletAddress={accountAddress}
-                  onViewMore={handleTabChange(TabIndex.Delegations)}
+                  onViewMore={handleTabChange(TabIndex.Delegations, undefined)}
                 />
               </Flex>
             )}
@@ -280,7 +299,10 @@ const AccountDetailsBody = ({
               address={accountAddress}
               scrollComponentId={tableHeaderId}
               refetchCount={refetchCounts}
-              onViewMore={handleTabChange(TabIndex.Txs)}
+              onViewMore={handleTabChange(
+                TabIndex.Txs,
+                tableCounts.txsCount ?? undefined
+              )}
             />
             {wasm.enabled && (
               <>
@@ -289,21 +311,30 @@ const AccountDetailsBody = ({
                   scrollComponentId={tableHeaderId}
                   totalData={tableCounts.codesCount ?? undefined}
                   refetchCount={refetchCounts}
-                  onViewMore={handleTabChange(TabIndex.Codes)}
+                  onViewMore={handleTabChange(
+                    TabIndex.Codes,
+                    tableCounts.codesCount ?? undefined
+                  )}
                 />
                 <InstantiatedContractsTable
                   walletAddress={accountAddress}
                   scrollComponentId={tableHeaderId}
                   totalData={tableCounts.contractsCount ?? undefined}
                   refetchCount={refetchCounts}
-                  onViewMore={handleTabChange(TabIndex.Contracts)}
+                  onViewMore={handleTabChange(
+                    TabIndex.Contracts,
+                    tableCounts.contractsCount ?? undefined
+                  )}
                 />
                 <AdminContractsTable
                   walletAddress={accountAddress}
                   scrollComponentId={tableHeaderId}
                   totalData={tableCounts.contractsAdminCount ?? undefined}
                   refetchCount={refetchCounts}
-                  onViewMore={handleTabChange(TabIndex.Admins)}
+                  onViewMore={handleTabChange(
+                    TabIndex.Admins,
+                    tableCounts.contractsAdminCount ?? undefined
+                  )}
                 />
               </>
             )}
@@ -314,14 +345,17 @@ const AccountDetailsBody = ({
                   totalCount={resourcesData?.totalCount}
                   resourcesByName={resourcesData?.groupedByName}
                   isLoading={isResourceLoading}
-                  onViewMore={handleTabChange(TabIndex.Resources)}
+                  onViewMore={handleTabChange(
+                    TabIndex.Resources,
+                    resourcesData?.groupedByOwner.length
+                  )}
                 />
                 <ModuleLists
                   totalCount={modulesData?.length}
                   selectedAddress={accountAddress}
                   modules={modulesData}
                   isLoading={isModulesLoading}
-                  onViewMore={handleTabChange(TabIndex.Modules)}
+                  onViewMore={handleTabChange(TabIndex.Modules, undefined)}
                 />
               </>
             )}
@@ -331,7 +365,10 @@ const AccountDetailsBody = ({
                 scrollComponentId={tableHeaderId}
                 totalData={tableCounts.proposalsCount ?? undefined}
                 refetchCount={refetchCounts}
-                onViewMore={handleTabChange(TabIndex.Proposals)}
+                onViewMore={handleTabChange(
+                  TabIndex.Proposals,
+                  tableCounts.proposalsCount ?? undefined
+                )}
               />
             )}
           </TabPanel>