+++ src/lib/pages/contract-details/components/contract-states/index.tsx
@@ -165,7 +165,8 @@ export const ContractStates = ({ contractAddress }: ContractStatesProps) => {
           const newVal = e.target.value;
           setKeyword(newVal);
         }}
-        action="contract-states-search"
+        size={{ base: "md", md: "lg" }}
+        amptrackSection="contract-states-search"
       />
 
       {/* State List */}