+++ src/lib/pages/past-txs/index.tsx
@@ -1,10 +1,4 @@
-import {
-  Flex,
-  Heading,
-  Input,
-  InputGroup,
-  InputRightElement,
-} from "@chakra-ui/react";
+import { Flex, Heading } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import type { ChangeEvent } from "react";
 import { useEffect, useMemo } from "react";
@@ -12,7 +6,7 @@ import { useForm } from "react-hook-form";
 
 import { AmpEvent, track } from "lib/amplitude";
 import { useCurrentChain, useWasmConfig } from "lib/app-provider";
-import { CustomIcon } from "lib/components/icon";
+import InputWithIcon from "lib/components/InputWithIcon";
 import PageContainer from "lib/components/PageContainer";
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
@@ -129,19 +123,15 @@ const PastTxs = () => {
         Past Transactions
       </Heading>
       <Flex my={8} gap={3}>
-        <InputGroup>
-          <Input
-            value={pastTxsState.search}
-            onChange={handleOnSearchChange}
-            placeholder={`Search with Transaction Hash${
-              wasm.enabled ? " or Contract Address" : ""
-            }`}
-            h="full"
-          />
-          <InputRightElement pointerEvents="none" h="full" mr={1}>
-            <CustomIcon name="search" color="gray.600" />
-          </InputRightElement>
-        </InputGroup>
+        <InputWithIcon
+          placeholder={`Search with Transaction Hash${
+            wasm.enabled ? " or Contract Address" : ""
+          }`}
+          value={pastTxsState.search}
+          onChange={handleOnSearchChange}
+          size={{ base: "md", md: "lg" }}
+          amptrackSection="past-txs-search"
+        />
         <Flex gap={3}>
           <TxRelationSelection
             value={pastTxsState.isSigner}